Move models to own subpackage
