+++ ExamApp/src/main/java/com/OBJ2100/ExamApp/db/DatabaseInterface.java
@@ -3,6 +3,9 @@ package com.OBJ2100.ExamApp.db;
 import java.sql.SQLException;
 import java.util.List;
 
+import com.OBJ2100.ExamApp.db.models.Customer;
+import com.OBJ2100.ExamApp.db.models.Employee;
+
 /**
  * Implementation of the methods for establishing database connection
  */
@@ -23,10 +26,10 @@ public interface DatabaseInterface {
 	List<Employee> getEmployees() throws SQLException;
 	
 	//add employee with name
-	void addEmployee(String firtsName, String LastName) throws SQLException;
+	void addEmployee(String firstName, String lastName) throws SQLException;
 	
 	//add employee with all details
-	void addEmployee(String firtsName, String LastName, String department, String email, double salary) throws SQLException;
+	void addEmployee(String firstName, String lastName, String department, String email, double salary) throws SQLException;
 	
 	// Customer
 	List<Customer> getCustomers() throws SQLException;
