+++ ExamApp/src/main/java/com/OBJ2100/ExamApp/db/models/Customer.java
@@ -1,4 +1,4 @@
-package com.OBJ2100.ExamApp.db;
+package com.OBJ2100.ExamApp.db.models;
 
 public class Customer {
 	private int customerNumber, salesRepEmployeeNumber;
