+++ ExamApp/src/main/java/com/OBJ2100/ExamApp/db/models/Employee.java
@@ -1,4 +1,4 @@
-package com.OBJ2100.ExamApp.db;
+package com.OBJ2100.ExamApp.db.models;
 
 public class Employee {
 	private int employeeNumber, reportsTo;
