+++ ExamApp/src/main/java/com/OBJ2100/ExamApp/db/models/Order.java
@@ -1,4 +1,4 @@
-package com.OBJ2100.ExamApp.db;
+package com.OBJ2100.ExamApp.db.models;
 
 import java.util.Date;
